% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth.R
\name{ache_ht}
\alias{ache_ht}
\title{Ache heights}
\usage{
ache_ht(age, sex)
}
\arguments{
\item{age}{Age in years}

\item{sex}{0: female, 1: male}
}
\value{
Height in centimeters
}
\description{
Ache height by age and sex
}
\details{
Based on a GAM fit to data points extracted from a plot in Ache Life History
}
\examples{
\dontrun{
if(interactive()){
 ache_ht(c(5, 10, 15), 1) # Male
 }
}
}
