% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth.R
\name{ache_wt}
\alias{ache_wt}
\title{Ache weights}
\usage{
ache_wt(age, sex, pregnant = 0)
}
\arguments{
\item{age}{Age in years}

\item{sex}{0: female, 1: male}

\item{pregnant}{Pregnant: 1, Default: 0}
}
\value{
Weight in kilograms
}
\description{
Ache weight by age, sex, and. pregnancy status
}
\details{
Uses parameters of the JPPS model fit to Ache data in https://onlinelibrary.wiley.com/doi/epdf/10.1002/ajpa.20306
}
\examples{
\dontrun{
if(interactive()){
 ache_wt(c(5, 10, 15), 1) # Male
 }
}
}
