% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth.R
\name{hadza_ht}
\alias{hadza_ht}
\title{Hadza heights}
\usage{
hadza_ht(age, sex)
}
\arguments{
\item{age}{Age in years}

\item{sex}{0: female, 1: male}
}
\value{
Height in centimeters
}
\description{
Height of Hadza by age and sex
}
\details{
Predictions from a GAM model fit to data digitized from a scatterplot in the growth SI to Demography And Evolutionary Ecology Of Hadza (Blurton-Jones).
}
\examples{
\dontrun{
if(interactive()){
 hadza_ht(c(5, 10, 15), 0) # Female
 }
}
}
