% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth.R
\name{hadza_wt}
\alias{hadza_wt}
\title{Hadza weights}
\usage{
hadza_wt(age, sex, pregnant = 0)
}
\arguments{
\item{age}{Age in years}

\item{sex}{0: female, 1: male}

\item{pregnant}{Pregnant: 1, Default: 0}
}
\value{
Weight in kilograms
}
\description{
Weight of Hadza by age and sex
}
\details{
Predictions from a GAM model fit to data digitized from a scatterplot in the growth SI to Demography And Evolutionary Ecology Of Hadza (Blurton-Jones).
}
\examples{
\dontrun{
if(interactive()){
 hadza_wt(c(25, 30, 35), 0, 1) # Pregnant women
 }
}
}
