% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth.R
\name{hg_height}
\alias{hg_height}
\title{Hunter-gatherer heights}
\usage{
hg_height(age, sex, group)
}
\arguments{
\item{age}{Age in years}

\item{sex}{0: female, 1: male}

\item{group}{One of 'ache', 'kung', 'hadza', 'avg'}
}
\value{
Height in centimeters
}
\description{
Hunter gather height by age and sex
}
\details{
'avg': height averaged across the three populations
}
\examples{
\dontrun{
if(interactive()){
 hg_weight(c(5, 10, 15), 1, 0, "avg") # The heights of 3 "average" hunter-gatherer boys
 }
}
}
