% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifecourse.R
\name{hg_lifecourse2}
\alias{hg_lifecourse2}
\title{hg_lifecourse2}
\usage{
hg_lifecourse2(
  SRB = 1.05,
  afb = 20,
  max_age = 80,
  group = "avg",
  e0_f = 35,
  e0_m = 30,
  TEE_prop_m = 2,
  alpha_m = 0.5,
  b1_m = 0.25,
  age50_m = 15,
  TEE_prop_f = 1,
  alpha_f = 0.5,
  b1_f = 0.25,
  age50_f = 15,
  ...
)
}
\arguments{
\item{SRB}{Sex ratio at birth, Default: 1.05}

\item{max_age}{Maximum age, Default: 80}

\item{group}{one of 'ache', 'hadza', 'kung', 'avg', Default: 'avg'}

\item{e0_f}{Expected female lifespan at birth, Default: 35}

\item{e0_m}{Expected male lifespan at birth, Default: 30}

\item{TEE_prop_m}{Male productivity as a proportion of average adult TEE, Default: 2}

\item{alpha_m}{The relative contribution of skill and strength to productivity (0: all skill; 1: all strength), Default: 0.5}

\item{b1_m}{Male rate of skill acquisition, Default: 0.25}

\item{age50_m}{Age at which male skill is 50\% of maximum, Default: 15}

\item{TEE_prop_f}{Female productivity as a proportion of average adult TEE, Default: 1}

\item{alpha_f}{The relative contribution of skill and strength to productivity (0: all skill; 1: all strength), Default: 0.5}

\item{b1_f}{Female rate of skill acquisition, Default: 0.25}

\item{age50_f}{Age at which female skill is 50\% of maximum, Default: 15}
}
\value{
A tibble (data frame)
}
\description{
Similar to hg_lifecourse but without creating families
}
\details{
One row for each age from birth to death
}
\examples{
\dontrun{
if(interactive()){
 #hg_lifecourse2()
 }
}
}
