% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/production.R
\name{hg_productivity}
\alias{hg_productivity}
\title{Hunter-gatherer productivity}
\usage{
hg_productivity(
  age,
  sex,
  TEE_prop = 2,
  alpha = 1,
  b1 = 0.25,
  age50 = 15,
  group = "avg"
)
}
\arguments{
\item{age}{Age in years}

\item{sex}{0: female; 1: male}

\item{TEE_prop}{Proportion of adult TEE, Default: 2}

\item{alpha}{Relative importance of strength vs skill. 0: all skill; 1: all strength, Default: 1}

\item{b1}{Rate of skill acquisition, Default: 0.25}

\item{age50}{Age at which skill is 50\% of adult Default: 15}

\item{group}{character 'ache', 'hadza', 'kung', 'avg'. Default: 'avg'}
}
\value{
Productivity as a proportion of adult TEE
}
\description{
Productivity as a function of age, sex, and group
}
\details{
\code{TEE_prop * hg_strength^alpha * hg_skill^(1-alpha)}
}
\examples{
\dontrun{
if(interactive()){
 hg_productivity(c(5, 15, 25), 0, TEE_prop = 1.5, alpha = 0.5, age50 = 15)
 }
}
}
