% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/production.R
\name{hg_productivity_avg}
\alias{hg_productivity_avg}
\title{Average hunter-gatherer productivity}
\usage{
hg_productivity_avg(
  age,
  TEE_prop_f,
  alpha_f,
  b1_f,
  age50_f,
  TEE_prop_m,
  alpha_m,
  b1_m,
  age50_m
)
}
\arguments{
\item{age}{Age in years}

\item{TEE_prop_f}{Proportion of adult female TEE}

\item{alpha_f}{Relative importance of female strength vs skill. 0: all skill; 1: all strength}

\item{b1_f}{Rate of female skill acquisition}

\item{age50_f}{Age at which female skill is 50\% of adult}

\item{TEE_prop_m}{Proportion of adult male TEE}

\item{alpha_m}{Relative importance of male strength vs skill. 0: all skill; 1: all strength}

\item{b1_m}{Rate of male skill acquisition}

\item{age50_m}{Age at which male skill is 50\% of adult}
}
\value{
Productivity as a proportion of adult TEE
}
\description{
Productivity as a function of age and group averaged across sex
}
\details{
\code{TEE_prop * hg_strength^alpha * hg_skill^(1-alpha)}
}
\examples{
\dontrun{
if(interactive()){
 hg_productivity(c(5, 15, 25), TEE_prop_f = 1.5, TEE_prop_m = 2.5, alpha = 0.5, age50 = 15)
 }
}
}
