% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth.R
\name{hg_weight}
\alias{hg_weight}
\title{Hunter-gatherer weights}
\usage{
hg_weight(age, sex, pregnant, group)
}
\arguments{
\item{age}{Age in years}

\item{sex}{0: female, 1: male}

\item{pregnant}{Pregnant: 1, Default: 0}

\item{group}{One of 'ache', 'kung', 'hadza', 'avg'}
}
\value{
Weight in kilograms
}
\description{
Hunter gather weight by age and sex
}
\details{
'avg': weight averaged across the three populations
}
\examples{
\dontrun{
if(interactive()){
 hg_weight(c(5, 10, 15), 1, 0, "avg") # The weights of 3 "average" hunter-gatherer boys
 }
}
}
