% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth.R
\name{jpps}
\alias{jpps}
\title{JPPS human growth model}
\usage{
jpps(age, params)
}
\arguments{
\item{age}{Age in years}

\item{params}{A numeric vector c(B, C1, C2, C3, D1, D2, D3)}
}
\value{
Weight in kilograms
}
\description{
JPPS 7 parameter model of human growth curves
}
\details{
Jolicoeur, Pontier, Pernin, and  Sempe (1988) A Lifetime Asymptotic Growth Curve for Human Height
}
\examples{
\dontrun{
if(interactive()){
 Ache_female <- c(B=55.63, C1=15.60, C2=3.58, C3=0.67, D1=13.60, D2=14.24, D3=26.95)
 jpps(25, Ache_female)
 }
}
}
