% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/digitized.R
\name{kaplan2000_hadza_male}
\alias{kaplan2000_hadza_male}
\title{Hadza male productivity}
\usage{
kaplan2000_hadza_male(age)
}
\arguments{
\item{age}{Age in years}
}
\value{
kcals
}
\description{
Hadza male productivity (kcals/day) digitized from plot in Kaplan et al. 2000
}
\details{
Interpolation of digitized from plot in Kaplan et al. 2000
}
\examples{
\dontrun{
if(interactive()){
 kaplan2000_hadza_male(5:60)
 }
}
}
