% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/production.R
\name{koster2020skill}
\alias{koster2020skill}
\title{Skill Koster et al 2020}
\usage{
koster2020skill(age)
}
\arguments{
\item{age}{Age in years}
}
\value{
Relative skill level
}
\description{
Relative skill as a function of age from Koster et al. (2020)
}
\details{
See Koster et al. (2020)
}
\examples{
\dontrun{
if(interactive()){
 koster2020skill(0:80)
 }
}
}
