% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth.R
\name{kung_ht}
\alias{kung_ht}
\title{!Kung heights}
\usage{
kung_ht(age, sex)
}
\arguments{
\item{age}{Age in years}

\item{sex}{0: female, 1: male}
}
\value{
Height in centimeters
}
\description{
Height in cm of !Kung by age and sex
}
\details{
Predictions of GAM model fit to Howell data
}
\examples{
\dontrun{
if(interactive()){
 kung_wt(c(5, 10, 15), 0) # Female
 }
}
}
