% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth.R
\name{kung_wt}
\alias{kung_wt}
\title{!Kung weights}
\usage{
kung_wt(age, sex, pregnant = 0)
}
\arguments{
\item{age}{Age in years}

\item{sex}{0: female, 1: male}

\item{pregnant}{Pregnant = 1, Default: 0}
}
\value{
Weight in kilograms
}
\description{
Weight in kilograms of !Kung by age, sex, and pregnancy
}
\details{
Predictions of GAM model fit to Howell data
}
\examples{
\dontrun{
if(interactive()){
 kung_wt(c(5, 10, 15), 0, 0) # Female
 }
}
}
