% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortality.R
\name{lifetable_UN}
\alias{lifetable_UN}
\title{UN Model Life Table}
\usage{
lifetable_UN(e0, sex, SRB = 1.05)
}
\arguments{
\item{e0}{Life expectancy at birth}

\item{sex}{character "Female" or "Male"}

\item{SRB}{Sex ratio at birth, Default: 1}
}
\value{
a UN Model Life Table
}
\description{
Returns a UN Model Life Table for a given life expectancy, sex, and SRB
}
\details{
Code from Gaddy et al. 2025, DOI: 10.1073/pnas.2508091122; https://www.pnas.org/doi/10.1073/pnas.2508091122
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
