% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortality.R
\name{lt_col_interpolate}
\alias{lt_col_interpolate}
\title{Interpolate life table quantities}
\usage{
lt_col_interpolate(col, e0, sex)
}
\arguments{
\item{col}{The life table quantity to interpolate}

\item{e0}{Life expectancy at birth}

\item{sex}{character "Female" or "Male"}
}
\value{
The interpolated vector
}
\description{
Interpolates lifetable quantities for fractional life expectancies, e.g, e0 = 30.7
}
\details{
Code from Gaddy et al. 2025, DOI: 10.1073/pnas.2508091122; https://www.pnas.org/doi/10.1073/pnas.2508091122
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
