% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortality.R
\name{LT_from_mx}
\alias{LT_from_mx}
\title{Life table from age-specific mortality rate}
\usage{
LT_from_mx(mx, ax, sex, SRB)
}
\arguments{
\item{mx}{age-specific mortality RATE (age x to x+n); ~ observed deaths / mid-interval population}

\item{ax}{number of years lived between x and x + n for those who died in the interval; ax will usually close to 0.5, except for infants and the very elderly}

\item{sex}{character "Female" or "Male"}

\item{SRB}{Sex ratio at birth (e.g., 1.05)}
}
\value{
A data frame with x, ax, mx, qx, px, lx, dx, Lx, Tx, ex
}
\description{
Build a life table from mx, ax, sex, and SRB
}
\details{
Code from Gaddy et al. 2025, DOI: 10.1073/pnas.2508091122; https://www.pnas.org/doi/10.1073/pnas.2508091122
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
