% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TEE.R
\name{TEE2}
\alias{TEE2}
\title{TEE from age, sex, group, elevation}
\usage{
TEE2(age, sex, group, pregnant = 0, elevation = 1000)
}
\arguments{
\item{age}{Age in years}

\item{sex}{0: female; 1: male}

\item{group}{character "ache", "hadza", "kung", "avg"}

\item{pregnant}{0: No; 1: Yes, Default: 0}

\item{elevation}{Elevation in meters, Default: 1000}
}
\value{
Total energy expenditure in kcals/day
}
\description{
Total energy expenditure (TEE) from hunter-gatherer age, sex, group, pregnancy status, and elevation
}
\details{
Formula for infants (age 0-1) from Pontzer et al. (2021); otherwise from Bajunaid et al. (2025) https://www.nature.com/articles/s43016-024-01089-5; heights and weights from the specified group.
}
\examples{
\dontrun{
if(interactive()){
 TEE2(25, 1, "hadza") # TEE of 25 year old Hadza male
 }
}
}
