% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TEE.R
\name{TEE_Bajunaid}
\alias{TEE_Bajunaid}
\title{TEE from body mass, height, age, sex, elevation}
\usage{
TEE_Bajunaid(bodymass, height, age, sex, elevation = 1000)
}
\arguments{
\item{bodymass}{body mass in kilograms}

\item{height}{Height in centimeters}

\item{age}{Age in years}

\item{sex}{0: female; 1: male}

\item{elevation}{elevation of residence in meters, Default: 1000}
}
\value{
Total energy expenditure (TEE) in kcals/day
}
\description{
Total energy expenditure (TEE) from body mass, height, age, sex, and elevation
}
\details{
Formula from Bajunaid et al. (2025) https://www.nature.com/articles/s43016-024-01089-5
}
\examples{
\dontrun{
if(interactive()){
 TEE_Bajunaid(50, 150, 25, 0)
 }
}
}
