% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TEE.R
\name{TEE_Ponzter}
\alias{TEE_Ponzter}
\title{TEE from body mass, age, and sex}
\usage{
TEE_Ponzter(bodymass, age, sex)
}
\arguments{
\item{bodymass}{Body mass in kilograms}

\item{age}{Age in years}

\item{sex}{0: female; 1: male}
}
\value{
otal energy expenditure (TEE) in kcals/day
}
\description{
Total energy expenditure (TEE) from body mass, age, and sex
}
\details{
Formula from Pontzer et al. (2021). Science 373, 808–812, Table S2
}
\examples{
\dontrun{
if(interactive()){
 TEE_Ponzter(50, 25, 1) # A 25 yr old male weighing 50 kg
 }
}
}
