% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TEE.R
\name{TEElog}
\alias{TEElog}
\title{TEE from body mass, age, sex, and percent fat free mass}
\usage{
TEElog(bodymass, age, sex, FFMpct = 0.85)
}
\arguments{
\item{bodymass}{Body mass in kilograms}

\item{age}{Age in years}

\item{sex}{0: female; 1: male}

\item{FFMpct}{Percent fat free mass, Default: 0.85}
}
\value{
Total energy expenditure (TEE) in kcals/day
}
\description{
TEE from body mass, age, sex, and percent fat free mass
}
\details{
Formula from Pontzer et al. (2021). Science 373, 808–812, Table S2
}
\examples{
\dontrun{
if(interactive()){
 TEElog(50, 25, 0)
 }
}
}
