% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifecourse.R
\name{hg_lifecourse}
\alias{hg_lifecourse}
\title{Hunter-gatherer lifecourse}
\usage{
hg_lifecourse(
  age_gap = 5,
  SRB = 1.05,
  afb = 18,
  max_age = 80,
  IBI = 3,
  preg_cost = 0.1,
  menopause_age = 80,
  group = "avg",
  e0_f = 35,
  e0_m = 30,
  TEE_prop_m = 2,
  alpha_m = 0.5,
  b1_m = 0.25,
  age50_m = 15,
  TEE_prop_f = 1,
  alpha_f = 0.5,
  b1_f = 0.25,
  age50_f = 15
)
}
\arguments{
\item{age_gap}{The age gap at marriage (e.g., 5: husband is 5 years older than the wife)}

\item{SRB}{Sex ratio at birth, Default: 1.05}

\item{afb}{Age at first birth, Default: 18}

\item{max_age}{Maximum human lifespan in years, Default: 80}

\item{IBI}{Interbirth interval (currently cannot be changed), Default: 3}

\item{preg_cost}{The cost of pregnancy (currently not used), Default: 0.1}

\item{menopause_age}{The age at menopause in years (80: no menopause), Default: 80}

\item{group}{character, one of 'ache', 'hadza', 'kung', 'avg', Default: 'avg'}

\item{e0_f}{Female expected lifespan at birth, Default: 35}

\item{e0_m}{Male expected lifespan at birth , Default: 30}

\item{TEE_prop_m}{Male energy production as a proportion of adult male TEE, Default: 2}

\item{alpha_m}{The relative contribution of skill and strength to productivity (0: all skill; 1: all strength), Default: 0.5}

\item{b1_m}{The rate of skill acquisition (0.2: slow; 0.4: fast), Default: 0.25}

\item{age50_m}{The age in years at which male skill reaches 50\% of maximum, Default: 15}

\item{TEE_prop_f}{Female energy production as a proportion of adult male (yes, male) TEE, Default: 1}

\item{alpha_f}{The relative contribution of skill and strength to productivity (0: all skill; 1: all strength), Default: 0.5}

\item{b1_f}{The rate of skill acquisition (0.2: slow; 0.4: fast), Default: 0.25}

\item{age50_f}{The age in years at which female skill reaches 50\% of maximum, Default: 15}
}
\value{
A tibble (data frame)
}
\description{
Simulation of a nuclear family from the wife's age at first birth to the maximum human lifespan
}
\details{
The function returns a tibble (data frame) with these columns and one row
for each year of the wife's age:

\code{wife_age, husband_age, wife_num, husband_num,
wife_survival, husband_survival, husband_wife_ratio, wife_ex, husband_ex,
pregnancy, births, fertility, child_age, child_index, girl_survival, boy_survival,
child_survival, wifeTEE, husbandTEE, childTEE, wife_production, husband_production,
child_production, resident_children, total_child_consumption, total_child_production,
family_size, family_consumption, family_production, energy_balance.}

The age at first birth (afb) is also the age of child independence.

The resident_children, total_child_consumption, total_child_production values are for
resident children only, and therefore are computed for the wife's latest birth
to her oldest resident child using the internal function \code{sum_resident_children}.
Additionally, whereas the other child values are conditional on the child being born,
these values are conditional on the wife surviving to give birth.
}
\examples{
\dontrun{
if(interactive()){
 hg_lifecourse()
 }
}
}
