% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/production.R
\name{hg_productivity_avg}
\alias{hg_productivity_avg}
\title{Average hunter-gatherer productivity}
\usage{
hg_productivity_avg(
  age,
  TEE_prop_f,
  alpha_f,
  b1_f,
  age50_f,
  TEE_prop_m,
  alpha_m,
  b1_m,
  age50_m
)
}
\arguments{
\item{age}{Age in years}

\item{sex}{0: female; 1: male}

\item{TEE_prop}{Proportion of adult TEE, Default: 2}

\item{alpha}{Relative importance of strength vs skill. 0: all skill; 1: all strength, Default: 1}

\item{b1}{Rate of skill acquisition, Default: 0.25}

\item{age50}{Age at which skill is 50\% of adult}

\item{group}{character 'ache', 'hadza', 'kung', 'avg'. Default: 'avg'}
}
\value{
Productivity as a proportion of adult TEE
}
\description{
Productivity as a function of age and group averaged across sex
}
\details{
\code{TEE_prop * hg_strength^alpha * hg_skill^(1-alpha)}
}
\examples{
\dontrun{
if(interactive()){
 hg_productivity(c(5, 15, 25), TEE_prop_f = 1.5, TEE_prop_m = 2.5, alpha = 0.5, age50 = 15)
 }
}
}
