% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/production.R
\name{hg_skill}
\alias{hg_skill}
\title{Hunter-gatherer skill}
\usage{
hg_skill(age, b1, age50 = 15, b2 = -0.15, max_age = 85)
}
\arguments{
\item{age}{Age in years}

\item{b1}{Rate of skill acquisition, e.g., 0.15 is slow, 0.40 is fast}

\item{age50}{Age when skill is 50\% of maximum, Default: 15}

\item{b2}{Rate of skill decline in old age, Default: -0.15}

\item{max_age}{Maximum age, Default: 85}
}
\value{
Skill level between 0 and 1
}
\description{
Skill given age, rate of acquisition, and age of 50\% level
}
\details{
Skill is computed as \code{ifelse(age < 3, 0, ( 1/(1 + exp(-b1 * (age - age50)))  ) * (1 - exp(b2 * (max_age-age))))}
}
\examples{
\dontrun{
if(interactive()){
 hg_skill(c(5, 10, 15), b1 = 0.25)
 }
}
}
