% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortality.R
\name{lifetable}
\alias{lifetable}
\title{Life tables for hunter-gatherer groups}
\usage{
lifetable(group, sex, e0 = NULL)
}
\arguments{
\item{group}{character "ache", "hadza", "kung", "avg"}

\item{sex}{0: female; 1: male}

\item{e0}{Life expectancy at birth (years), Default: NULL}
}
\value{
life table
}
\description{
Returns a life table for Ache, Hadza, !Kung, or average
}
\details{
Ache life table from Ache Life History; Hadza life table from Blurton-Jones (2016) Demography and evolutionary ecology of Hadza hunter-gatherers; !Kung and average life tables use code from Code from Gaddy et al. 2025, DOI: 10.1073/pnas.2508091122; https://www.pnas.org/doi/10.1073/pnas.2508091122
}
\examples{
\dontrun{
if(interactive()){
 lifetable("hadza", 0, e0 = 35) # Female life table for e0 = 35 years
 }
}
}
